+++ libs/dh/shared/feature-flags/src/index.ts
@@ -14,6 +14,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { type DhFeatureFlag } from './lib/feature-flags';
-export * from './lib/feature-flags.service';
-export * from './lib/feature-flag.directive';
+export { type DhFeatureFlag } from './lib/dh-feature-flags';
+export * from './lib/dh-feature-flags.service';
+export * from './lib/dh-feature-flag.directive';
